Add files via upload
This code requires some more tweaking, as demonstrated in param.py, some modules needs to be included.

Running this code won't fetch anything because initial steps are needed to be performed like starting kmd process and algod process
This code is intended to just create a wallet and an account associated with it. Give back the detail of the created account to the user like mnemoice phrase, key, id, etc.

This week I'll add up the code to start the kmd process and algod process. Meanwhile, please do look for any changes that should be made to the code.
The initialization of process is shown in param.py code. https://github.com/algorand/py-algorand-sdk/blob/master/params.py